+++ x/attribute/keeper/keeper_legacy_test.go
@@ -2,6 +2,7 @@ package keeper_test
 
 import (
 	"fmt"
+	"strings"
 	"testing"
 
 	tmproto "github.com/tendermint/tendermint/proto/tendermint/types"
@@ -17,7 +18,7 @@ import (
 	"github.com/stretchr/testify/suite"
 )
 
-type KeeperTestSuite struct {
+type KeeperLegacyTestSuite struct {
 	suite.Suite
 
 	app *app.App
@@ -32,11 +33,11 @@ type KeeperTestSuite struct {
 	user2Addr sdk.AccAddress
 }
 
-func TestKeeperTestSuite(t *testing.T) {
-	suite.Run(t, new(KeeperTestSuite))
+func TestKeeperLegacyTestSuite(t *testing.T) {
+	suite.Run(t, new(KeeperLegacyTestSuite))
 }
 
-func (s *KeeperTestSuite) SetupTest() {
+func (s *KeeperLegacyTestSuite) SetupTest() {
 	app := simapp.Setup(false)
 	ctx := app.BaseApp.NewContext(false, tmproto.Header{})
 	s.app = app
@@ -52,7 +53,7 @@ func (s *KeeperTestSuite) SetupTest() {
 
 	var nameData nametypes.GenesisState
 	nameData.Bindings = append(nameData.Bindings, nametypes.NewNameRecord("attribute", s.user1Addr, false))
-	nameData.Bindings = append(nameData.Bindings, nametypes.NewNameRecord("example.attribute", s.user1Addr, false))
+	nameData.Bindings = append(nameData.Bindings, nametypes.NewNameRecord("old.attribute", s.user1Addr, false))
 	nameData.Params.AllowUnrestrictedNames = false
 	nameData.Params.MaxNameLevels = 3
 	nameData.Params.MinSegmentLength = 3
@@ -64,9 +65,20 @@ func (s *KeeperTestSuite) SetupTest() {
 	params.MaxValueLength = 10
 	app.AttributeKeeper.SetParams(ctx, params)
 	s.app.AccountKeeper.SetAccount(s.ctx, s.app.AccountKeeper.NewAccountWithAddress(s.ctx, s.user1Addr))
+
+	var attributeData types.GenesisState
+	attributeData.Attributes = append(attributeData.Attributes, types.Attribute{
+		Name:          "old.attribute",
+		Value:         []byte("0123456789"),
+		Address:       s.user1,
+		AttributeType: types.AttributeType_String,
+	})
+
+	InitGenesisLegacy(s.ctx, &attributeData, s.app)
+	s.app.AttributeKeeper.ConvertLegacyAmino(s.ctx)
 }
 
-func (s *KeeperTestSuite) TestSetAttribute() {
+func (s *KeeperLegacyTestSuite) TestSetAttribute() {
 
 	cases := map[string]struct {
 		attr      types.Attribute
@@ -77,7 +89,7 @@ func (s *KeeperTestSuite) TestSetAttribute() {
 	}{
 		"should successfully add attribute": {
 			attr: types.Attribute{
-				Name:          "example.attribute",
+				Name:          "old.attribute",
 				Value:         []byte("0123456789"),
 				Address:       s.user1,
 				AttributeType: types.AttributeType_String,
@@ -164,15 +176,7 @@ func (s *KeeperTestSuite) TestSetAttribute() {
 
 }
 
-func (s *KeeperTestSuite) TestDeleteAttribute() {
-
-	attr := types.Attribute{
-		Name:          "example.attribute",
-		Value:         []byte("0123456789"),
-		Address:       s.user1,
-		AttributeType: types.AttributeType_String,
-	}
-	s.app.AttributeKeeper.SetAttribute(s.ctx, s.user1Addr, attr, s.user1Addr)
+func (s *KeeperLegacyTestSuite) TestDeleteAttribute() {
 
 	cases := map[string]struct {
 		name      string
@@ -182,7 +186,7 @@ func (s *KeeperTestSuite) TestDeleteAttribute() {
 		errorMsg  string
 	}{
 		"should fail to delete, cant resolve name wrong owner": {
-			name:      "example.attribute",
+			name:      "old.example.attribute",
 			accAddr:   s.user1Addr,
 			ownerAddr: s.user2Addr,
 			wantErr:   true,
@@ -196,7 +200,7 @@ func (s *KeeperTestSuite) TestDeleteAttribute() {
 			errorMsg:  fmt.Sprintf("\"dne\" does not resolve to address \"%s\"", s.user1Addr),
 		},
 		"should successfully delete attribute": {
-			name:      "example.attribute",
+			name:      "old.attribute",
 			accAddr:   s.user1Addr,
 			ownerAddr: s.user1Addr,
 			wantErr:   false,
@@ -219,19 +223,18 @@ func (s *KeeperTestSuite) TestDeleteAttribute() {
 
 }
 
-func (s *KeeperTestSuite) TestGetAllAttributes() {
+func (s *KeeperLegacyTestSuite) TestGetAllAttributes() {
 
 	attributes, err := s.app.AttributeKeeper.GetAllAttributes(s.ctx, s.user1Addr)
 	s.NoError(err)
-	s.Equal(0, len(attributes))
+	s.Equal(1, len(attributes))
 
 	attr := types.Attribute{
-		Name:          "example.attribute",
+		Name:          "old.attribute",
 		Value:         []byte("0123456789"),
 		Address:       s.user1,
 		AttributeType: types.AttributeType_String,
 	}
-	s.app.AttributeKeeper.SetAttribute(s.ctx, s.user1Addr, attr, s.user1Addr)
 
 	attributes, err = s.app.AttributeKeeper.GetAllAttributes(s.ctx, s.user1Addr)
 	s.NoError(err)
@@ -240,20 +243,19 @@ func (s *KeeperTestSuite) TestGetAllAttributes() {
 	s.Equal(attr.Value, attributes[0].Value)
 }
 
-func (s *KeeperTestSuite) TestGetAttributesByName() {
+func (s *KeeperLegacyTestSuite) TestGetAttributesByName() {
 
 	attr := types.Attribute{
-		Name:          "example.attribute",
+		Name:          "old.attribute",
 		Value:         []byte("0123456789"),
 		Address:       s.user1,
 		AttributeType: types.AttributeType_String,
 	}
-	s.app.AttributeKeeper.SetAttribute(s.ctx, s.user1Addr, attr, s.user1Addr)
 
 	_, err := s.app.AttributeKeeper.GetAttributes(s.ctx, s.user1Addr, "blah")
 	s.Error(err)
 	s.Equal("no address bound to name", err.Error())
-	attributes, err := s.app.AttributeKeeper.GetAttributes(s.ctx, s.user1Addr, "example.attribute")
+	attributes, err := s.app.AttributeKeeper.GetAttributes(s.ctx, s.user1Addr, "old.attribute")
 	s.NoError(err)
 	s.Equal(1, len(attributes))
 	s.Equal(attr.Name, attributes[0].Name)
@@ -261,7 +263,7 @@ func (s *KeeperTestSuite) TestGetAttributesByName() {
 	s.Equal(attr.Value, attributes[0].Value)
 }
 
-func (s *KeeperTestSuite) TestInitGenesisAddingAttributes() {
+func (s *KeeperLegacyTestSuite) TestInitGenesisAddingAttributes() {
 	var attributeData types.GenesisState
 	attributeData.Attributes = append(attributeData.Attributes, types.Attribute{
 		Name:          "example.attribute",
@@ -270,7 +272,6 @@ func (s *KeeperTestSuite) TestInitGenesisAddingAttributes() {
 		AttributeType: types.AttributeType_String,
 	})
 	s.Assert().NotPanics(func() { s.app.AttributeKeeper.InitGenesis(s.ctx, &attributeData) })
-	s.Assert().NotPanics(func() { s.app.AttributeKeeper.ExportGenesis(s.ctx) })
 
 	attributeData.Attributes = append(attributeData.Attributes, types.Attribute{
 		Name:          "",
@@ -282,7 +283,7 @@ func (s *KeeperTestSuite) TestInitGenesisAddingAttributes() {
 	s.Assert().Panics(func() { s.app.AttributeKeeper.InitGenesis(s.ctx, &attributeData) })
 }
 
-func (s *KeeperTestSuite) TestIterateRecord() {
+func (s *KeeperLegacyTestSuite) TestIterateRecord() {
 	s.Run("iterate attribute's", func() {
 		attr := types.Attribute{
 			Name:          "example.attribute",
@@ -299,9 +300,57 @@ func (s *KeeperTestSuite) TestIterateRecord() {
 			return nil
 		}
 		// Collect and return genesis state.
-		err := s.app.AttributeKeeper.IterateRecords(s.ctx, types.AttributeKeyPrefix, appendToRecords)
+		err := s.app.AttributeKeeper.IterateRecords(s.ctx, types.AttributeKeyPrefixAmino, appendToRecords)
 		s.Require().NoError(err)
 		s.Require().Equal(1, len(records))
 	})
 
+}
+
+// InitGenesisLegacy creates the initial genesis state for the attribute module. ONLY FOR TESTING.
+func InitGenesisLegacy(ctx sdk.Context, data *types.GenesisState, app *app.App) {
+	if err := data.ValidateBasic(); err != nil {
+		panic(err)
+	}
+	for _, attr := range data.Attributes {
+		if err := importAttributeLegacy(ctx, attr, app); err != nil {
+			panic(err)
+		}
+	}
+}
+
+// A genesis helper that imports attribute state without owner checks.ONLY FOR TESTING.
+func importAttributeLegacy(ctx sdk.Context, attr types.Attribute, app *app.App) error {
+	// Ensure attribute is valid
+	if err := attr.ValidateBasic(); err != nil {
+		return err
+	}
+	// Attribute must have a valid, non-empty address to import
+	if strings.TrimSpace(attr.Address) == "" {
+		return fmt.Errorf("unable to import attribute with empty address")
+	}
+	acc, err := sdk.AccAddressFromBech32(attr.Address)
+	if err != nil {
+		return err
+	}
+	// Ensure name is stored in normalized format.
+	if attr.Name, err = app.NameKeeper.Normalize(ctx, attr.Name); err != nil {
+		return fmt.Errorf("unable to normalize attribute name \"%s\": %w", attr.Name, err)
+	}
+	// Store the sanitized account attribute
+	bz, err := types.ModuleCdc.MarshalBinaryBare(&attr)
+	if err != nil {
+		return err
+	}
+	key := accountAttributeKeyLegacy(acc, attr)
+	store := ctx.KVStore(app.GetKey(types.ModuleName))
+	store.Set(key, bz)
+	return nil
+}
+
+// ONLY FOR TESTING.
+func accountAttributeKeyLegacy(acc sdk.AccAddress, attr types.Attribute) []byte {
+	key := append(types.AttributeKeyPrefixAmino, acc.Bytes()...)
+	key = append(key, types.GetNameKeyBytes(attr.Name)...)
+	return append(key, attr.Hash()...)
 }