+++ CHANGELOG.md
@@ -40,6 +40,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * Remove deprecated ModuleCdc amino encoding from Metadata Locator records #187
 * Update Cosmos SDK to 0.42.3
 * Remove deprecated ModuleCdc amino encoding from name module #189
+* Remove deprecated ModuleCdc amino encoding from attribute module #188
 
 ### Features